requirements.txt
aiogram
python-dotenv
yt-dlp
moviepy